More Information Messages
Now, a lot more informational messages are sent, telling you things like turns not working, attacks being applied, counters countering, etc.

Other Important Things:
- Fixed a bunch of translations not properly working and added some new ones.
- Elo is now rounded in the scoreboard.
- The warp classes now actually filter if you can use turns or not.
- Fixed some boolean algebra stuff.
- Adjusted ranking to have the minimum limit of 250 instead of 100.
- Did some rebalancing of turns.
- Removed the requiresUsageGuide variable in turns, as it will automatically check for available translations now.

+++ src/main/kotlin/net/hectus/neobb/game/mode/DefaultGame.kt
@@ -87,7 +87,7 @@ class DefaultGame(world: World, bukkitPlayers: List<Player>, difficulty: GameDif
                 locale.component("scoreboard.rank", color = Colors.ACCENT)
                     .append(Rank.ofElo(player.databaseInfo.elo).toRankTranslations(locale)),
                 locale.component("scoreboard.elo", color = Colors.ACCENT)
-                    .append(Component.text(player.databaseInfo.elo, Colors.RESET)),
+                    .append(Component.text(player.databaseInfo.elo.toInt(), Colors.RESET)),
                 Component.empty(),
                 Component.text("NeoBB-" + NeoBB.VERSION + " (d" + Integer.toHexString(LocalDateTime.now().dayOfYear) + "h" + LocalDateTime.now().hour + ")", Colors.EXTRA),
                 Component.text("mc.hectus.net", Colors.LINK)
